Refactor crud.insert and create new file for other helper functions.
